feat: align with other application which separates each resource and typescript generator
